style: remove notes from code

+++ src/config/index.ts
@@ -22,7 +22,7 @@ const config = {
     elasticsearch: {
         indexInitRetries: 3,
         fuzziness: env.get('FUZZINESS').default('AUTO').asString(),
-        boost: env.get('BOOST').default(1.2).asInt(),
+        boost: env.get('BOOST').default(1.2).asFloat(),
         nodes: env.get('ELASTICSEARCH_HOSTS').required().asString().split(','),
         auth: {
             username: env.get('ELASTICSEARCH_USERNAME').default('').asString(),
